Move NTP to base class Switch to nuget reference and delete /lib copy of NTP client
