--- Lib/GuerrillaNtp/README.md
@@ -1,3 +1,0 @@
-Source code can be found here
-
-https://github.com/blehnen/GuerrillaNtp
