--- Lib/GuerrillaNtp/net10.0/GuerrillaNtp.deps.json
@@ -1,23 +1,0 @@
-{
-  "runtimeTarget": {
-    "name": ".NETCoreApp,Version=v10.0",
-    "signature": ""
-  },
-  "compilationOptions": {},
-  "targets": {
-    ".NETCoreApp,Version=v10.0": {
-      "GuerrillaNtp/1.0.0": {
-        "runtime": {
-          "GuerrillaNtp.dll": {}
-        }
-      }
-    }
-  },
-  "libraries": {
-    "GuerrillaNtp/1.0.0": {
-      "type": "project",
-      "serviceable": false,
-      "sha512": ""
-    }
-  }
-}
