+++ Source/DotNetWorkQueue.Transport.Redis/DotNetWorkQueue.Transport.Redis.csproj
@@ -64,28 +64,7 @@ https://github.com/blehnen/DotNetWorkQueue/blob/master/CHANGELOG.md </PackageRel
 
   <ItemGroup Condition=" '$(TargetFramework)' == 'net48' ">
     <Reference Include="Microsoft.CSharp" />
-    <Reference Include="GuerrillaNtp">
-      <HintPath>..\..\Lib\GuerrillaNtp\net48\GuerrillaNtp.dll</HintPath>
-    </Reference>
   </ItemGroup>
-
-  <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.0'">
-    <Reference Include="GuerrillaNtp">
-      <HintPath>..\..\Lib\GuerrillaNtp\netstandard2.0\GuerrillaNtp.dll</HintPath>
-    </Reference>
-  </ItemGroup>
-
-	<ItemGroup Condition="'$(TargetFramework)' == 'net10.0'">
-		<Reference Include="GuerrillaNtp">
-			<HintPath>..\..\Lib\GuerrillaNtp\net10.0\GuerrillaNtp.dll</HintPath>
-		</Reference>
-	</ItemGroup>
-
-	<ItemGroup Condition="'$(TargetFramework)' == 'net8.0'">
-		<Reference Include="GuerrillaNtp">
-			<HintPath>..\..\Lib\GuerrillaNtp\net8.0\GuerrillaNtp.dll</HintPath>
-		</Reference>
-	</ItemGroup>
 
   <ItemGroup>
     <None Include="README.md" Pack="true" PackagePath="\" />
