+++ Source/DotNetWorkQueue/DotNetWorkQueue.csproj
@@ -67,6 +67,7 @@ https://github.com/blehnen/DotNetWorkQueue/blob/master/CHANGELOG.md</PackageRele
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="GuerrillaNtp" Version="3.1.0" />
     <PackageReference Include="Microsoft.CSharp" Version="4.7.0" />
     <PackageReference Include="Microsoft.IO.RecyclableMemoryStream" Version="3.0.1" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.4" />
